+++ drizzle/0009_announced_at.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `clip_projects` ADD `announced_at` integer;